+++ packages/plugins/src/types.d.ts
@@ -1,0 +1,9 @@
+
+export interface DataLoader {
+  load: () => Promise<any>;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface DataSource {
+
+}